examples/subobc: apply rename migration(Library -> library)
