+++ examples/subobc/src/src_user/applications/driver_instances/di_mobc.c
@@ -6,10 +6,10 @@
 
 #include "di_mobc.h"
 #include <src_core/TlmCmd/packet_handler.h>
-#include <src_core/Library/print.h>
+#include <src_core/library/print.h>
 #include "../../Settings/port_config.h"
 #include "../../Settings/DriverSuper/driver_buffer_define.h"
-#include <src_core/Library/result.h>
+#include <src_core/library/result.h>
 
 static RESULT DI_MOBC_init_(void);
 static RESULT DI_MOBC_update_(void);
