+++ examples/subobc/src/src_user/library/vt100.c
@@ -1,7 +1,7 @@
 #pragma section REPRO
 #include "vt100.h"
 
-#include <src_core/Library/print.h>
+#include <src_core/library/print.h>
 
 void VT100_erase_down(void)
 {
