+++ examples/subobc/src/src_user/script/add_section/settings.pl
@@ -19,7 +19,7 @@ sub GetSetting {
 		'applications',
 		'Drivers',
 		'hal',
-		'Library',
+		'library',
 		'Settings',
 		'TlmCmd'
 	];
