+++ examples/subobc/sync_with_mobc_example.bat
@@ -8,7 +8,7 @@ echo.
 
 call :sync_dir ".\src\s2e_mockup\" "..\mobc\src\s2e_mockup\"
 call :sync_dir ".\src\src_user\BootLoader\" "..\mobc\src\src_user\BootLoader\"
-call :sync_dir ".\src\src_user\Library\" "..\mobc\src\src_user\Library\"
+call :sync_dir ".\src\src_user\library\" "..\mobc\src\src_user\library\"
 call :sync_dir ".\src\src_user\ReProgramming\" "..\mobc\src\src_user\ReProgramming\"
 call :sync_dir ".\src\src_user\script\" "..\mobc\src\src_user\script\"
 call :sync_dir ".\src\src_user\Settings\TlmCmd\Ccsds\" "..\mobc\src\src_user\Settings\TlmCmd\Ccsds\"
